update ui on logout
